# cmd/api/posts.go
package main

import (
	"net/http"
	"social/internal/models"

	"github.com/google/uuid"
)

// postPayload represents the payload to create a post
// swagger:model postPayload
type postPayload struct {
	// Title of the post
	// example: My first post
	Title string `json:"title" validate:"required,max=100" example:"My first post"`
	// Content of the post
	// example: Hello world!
	Content string `json:"content" validate:"required,max=1000" example:"Hello world!"`
	// Tags associated with the post
	// example: ["go","backend"]
	Tags []string `json:"tags" example:"go,backend"`
}

// createPostHandler godoc
//
//	@Summary		Create a post
//	@Description	Creates a new post
//	@Tags			Posts
//	@Accept			json
//	@Produce		json
//	@Param			request	body		postPayload	true	"Post payload"
//	@Success		201		{object}	DataResponsePost
//	@Failure		400		{object}	ErrorResponse
//	@Failure		500		{object}	ErrorResponse
//	@Router			/posts [post]
func (app *application) createPostHandler(w http.ResponseWriter, r *http.Request) {
	var payload postPayload
	err := readJSON(w, r, &payload)
	if err != nil {
		app.errorBadRequest(w, r, err)
		return
	}

	err = Validate.Struct(payload)
	if err != nil {
		app.errorBadRequest(w, r, err)
		return
	}
	postID, err := uuid.NewV7()
	if err != nil {
		app.errorServerError(w, r, err)
		return
	}
	// todo: get userID from request context
	userID := uuid.MustParse("b58e1f73-028f-4c17-b8ac-8a3b416c69fd")

	post := &models.Post{
		ID:      postID,
		Title:   payload.Title,
		Content: payload.Content,
		Tags:    payload.Tags,
		UserID:  userID,
	}
	err = app.models.Posts.Create(r.Context(), post)
	if err != nil {
		app.errorServerError(w, r, err)
		return
	}

	err = app.jsonResponse(w, http.StatusCreated, post)
	if err != nil {
		app.errorServerError(w, r, err)
	}
}

// getPostHandler godoc
//
//	@Summary		Get a post
//	@Description	Retrieves a post by ID
//	@Tags			Posts
//	@Produce		json
//	@Param			postID	path		string	true	"Post ID (UUID)"
//	@Success		200		{object}	DataResponsePost
//	@Failure		404		{object}	ErrorResponse
//	@Failure		500		{object}	ErrorResponse
//	@Router			/posts/{postID} [get]
func (app *application) getPostHandler(w http.ResponseWriter, r *http.Request) {
	post := getPostFromContext(r)
	comments, err := app.models.Comments.GetComments(r.Context(), post.ID)
	if err != nil {
		app.errorServerError(w, r, err)
		return
	}
	post.Comments = comments

	err = app.jsonResponse(w, http.StatusOK, post)
	if err != nil {
		app.errorServerError(w, r, err)
	}
}

// deletePostHandler godoc
//
//	@Summary		Delete a post
//	@Description	Deletes a post by ID
//	@Tags			Posts
//	@Param			postID	path	string	true	"Post ID (UUID)"
//	@Success		204		"No Content"
//	@Failure		404		{object}	ErrorResponse
//	@Failure		500		{object}	ErrorResponse
//	@Router			/posts/{postID} [delete]
func (app *application) deletePostHandler(w http.ResponseWriter, r *http.Request) {
	post := getPostFromContext(r)
	err := app.models.Posts.Delete(r.Context(), post.ID)
	if err != nil {
		app.errorServerError(w, r, err)
		return
	}

	err = app.jsonResponse(w, http.StatusNoContent, nil)
	if err != nil {
		app.errorServerError(w, r, err)
	}
}

// updatePostPayload represents the payload to update a post
// swagger:model updatePostPayload
type updatePostPayload struct {
	// New title of the post
	// example: Updated title
	Title *string `json:"title" validate:"omitempty,max=100" example:"Updated title"`
	// New content of the post
	// example: Updated content
	Content *string `json:"content" validate:"omitempty,max=1000" example:"Updated content"`
	// New tags of the post
	// example: ["go","api"]
	Tags []string `json:"tags" validate:"omitempty" example:"go,api"`
}

// updatePostHandler godoc
//
//	@Summary		Update a post
//	@Description	Updates a post by ID
//	@Tags			Posts
//	@Accept			json
//	@Produce		json
//	@Param			postID	path		string				true	"Post ID (UUID)"
//	@Param			request	body		updatePostPayload	true	"Update payload"
//	@Success		200		{object}	DataResponsePost
//	@Failure		400		{object}	ErrorResponse
//	@Failure		404		{object}	ErrorResponse
//	@Failure		500		{object}	ErrorResponse
//	@Router			/posts/{postID} [patch]
func (app *application) updatePostHandler(w http.ResponseWriter, r *http.Request) {
	post := getPostFromContext(r)

	var payload updatePostPayload
	err := readJSON(w, r, &payload)
	if err != nil {
		app.errorBadRequest(w, r, err)
	}

	err = Validate.Struct(payload)
	if err != nil {
		app.errorBadRequest(w, r, err)
		return
	}

	if payload.Title != nil {
		post.Title = *payload.Title
	}
	if payload.Content != nil {
		post.Content = *payload.Content
	}
	if payload.Tags != nil {
		post.Tags = payload.Tags
	}

	err = app.models.Posts.Update(r.Context(), post)
	if err != nil {
		app.errorServerError(w, r, err)
		return
	}

	err = app.jsonResponse(w, http.StatusOK, post)
	if err != nil {
		app.errorServerError(w, r, err)
	}
}