# docs/swagger.yaml
basePath: /v1
definitions:
  main.DataResponseComment:
    properties:
      data:
        $ref: '#/definitions/models.Comment'
    type: object
  main.DataResponseFeed:
    properties:
      data:
        items:
          $ref: '#/definitions/models.FeedPost'
        type: array
    type: object
  main.DataResponsePost:
    properties:
      data:
        $ref: '#/definitions/models.Post'
    type: object
  main.DataResponseUser:
    properties:
      data:
        $ref: '#/definitions/models.User'
    type: object
  main.ErrorResponse:
    properties:
      error:
        description: |-
          Message describing the error
          example: invalid request payload
        type: string
    type: object
  main.HealthResponse:
    properties:
      env:
        example: development
        type: string
      status:
        example: ok
        type: string
      version:
        example: 0.0.1
        type: string
    type: object
  main.commentPayload:
    properties:
      content:
        description: |-
          Content of the comment
          example: Nice post!
        example: Nice post!
        maxLength: 1000
        type: string
    required:
    - content
    type: object
  main.postPayload:
    properties:
      content:
        description: |-
          Content of the post
          example: Hello world!
        example: Hello world!
        maxLength: 1000
        type: string
      tags:
        description: |-
          Tags associated with the post
          example: ["go","backend"]
        example:
        - go
        - backend
        items:
          type: string
        type: array
      title:
        description: |-
          Title of the post
          example: My first post
        example: My first post
        maxLength: 100
        type: string
    required:
    - content
    - title
    type: object
  main.updatePostPayload:
    properties:
      content:
        description: |-
          New content of the post
          example: Updated content
        example: Updated content
        maxLength: 1000
        type: string
      tags:
        description: |-
          New tags of the post
          example: ["go","api"]
        example:
        - go
        - api
        items:
          type: string
        type: array
      title:
        description: |-
          New title of the post
          example: Updated title
        example: Updated title
        maxLength: 100
        type: string
    type: object
  models.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      post_id:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.FeedPost:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      comments_count:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      top_comment_content:
        type: string
      top_comment_user_id:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
      version:
        type: integer
    type: object
  models.Post:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
      version:
        type: integer
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Social API server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Social API
  version: "1.0"
paths:
  /health:
    get:
      description: Returns service health information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.HealthResponse'
      summary: Health check
      tags:
      - Health
  /posts:
    post:
      consumes:
      - application/json
      description: Creates a new post
      parameters:
      - description: Post payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.postPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.DataResponsePost'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Create a post
      tags:
      - Posts
  /posts/{postID}:
    delete:
      description: Deletes a post by ID
      parameters:
      - description: Post ID (UUID)
        in: path
        name: postID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Delete a post
      tags:
      - Posts
    get:
      description: Retrieves a post by ID
      parameters:
      - description: Post ID (UUID)
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.DataResponsePost'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get a post
      tags:
      - Posts
    patch:
      consumes:
      - application/json
      description: Updates a post by ID
      parameters:
      - description: Post ID (UUID)
        in: path
        name: postID
        required: true
        type: string
      - description: Update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.updatePostPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.DataResponsePost'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Update a post
      tags:
      - Posts
  /posts/{postID}/comments:
    post:
      consumes:
      - application/json
      description: Creates a new comment associated with the specified post
      parameters:
      - description: Post ID (UUID)
        in: path
        name: postID
        required: true
        type: string
      - description: Comment payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.commentPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.DataResponseComment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Create a comment for a post
      tags:
      - Comments
  /users/{userID}:
    get:
      description: Retrieves a user by ID
      parameters:
      - description: User ID (UUID)
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.DataResponseUser'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get a user
      tags:
      - Users
  /users/{userID}/follow:
    put:
      description: Follow the specified user
      parameters:
      - description: User ID (UUID)
        in: path
        name: userID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Follow a user
      tags:
      - Users
  /users/{userID}/unfollow:
    put:
      description: Unfollow the specified user
      parameters:
      - description: User ID (UUID)
        in: path
        name: userID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Unfollow a user
      tags:
      - Users
  /users/feed:
    get:
      description: Returns a paginated feed of posts from the user and followed users
      parameters:
      - description: Items per page
        in: query
        maximum: 20
        minimum: 1
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        minimum: 0
        name: offset
        type: integer
      - description: Sort order
        enum:
        - ASC
        - DESC
        in: query
        name: sort
        type: string
      - collectionFormat: csv
        description: Filter by tags (comma separated)
        in: query
        items:
          type: string
        name: tags
        type: array
      - description: Search in title/content
        in: query
        name: search
        type: string
      - description: From date RFC3339
        example: '"2024-01-02T15:04:05Z"'
        in: query
        name: from
        type: string
      - description: To date RFC3339
        example: '"2024-12-31T23:59:59Z"'
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.DataResponseFeed'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get user feed
      tags:
      - Feed
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"